birdspotter/requirements.txt
asgiref==3.3.1
attrs==20.3.0
certifi==2020.11.8
click==7.1.2
click-plugins==1.1.1
cligj==0.7.0
Django==3.1.3
django-crispy-forms==1.11.0
django-private-storage==2.2.2
django-widget-tweaks==1.4.8
django-health-check==3.16.2
django-mailbox==4.8.2
Fiona==1.8.17
GDAL==3.2.1
geopandas==0.8.1
munch==2.5.0
numpy==1.19.4
Pillow==8.1.0
plotly==4.13.0
pyproj==3.0.0.post1
python-dateutil==2.8.1
pytz==2020.4
retrying==1.3.3
Shapely==1.7.1
six==1.15.0
sqlparse==0.4.1
urllib3==1.26.2
python-dotenv==0.15.0
psutil==5.8.0
whitenoise==5.2.0